Add unsafe code analysis workflows/CalculateUnsafeCode
Adds a new workflow that can be called to calculate the amount of
unsafe code across different categories in a Rust codebase. It is
designed to push the results to a branch in the repository where
the data can be picked up by a badge in the repo. The count-unsafe
tool is used to actually count the amount of unsafe code.

- `CalculateUnsafeCode.yml` - Reusable workflow that can be called
  in a repo to generate the data.
- `analyze_unsafe_code.py` - Helper functions that are used in the
  workflow to transform the data into a format that's ready to be
  used by shields.io's dynamic JSON badge support.

Results are broken down into these categories.

- Overall: Combined percentage across all categories
- Functions: Percentage of functions marked as `unsafe`
- Expressions: Percentage of `unsafe` expressions and blocks
- Implementations: Percentage of `unsafe` trait implementations
- Traits: Percentage of `unsafe` trait definitions
- Methods: Percentage of methods marked as `unsafe`

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>