Add support for unsafe code reporting with cargo geiger
Provides an alternative to the count-unsafe wrapper script which
remains in `.github/scripts/analyze_unsafe_code.py` with a script
the analyzes results from `cargo-geiger`.

Both scripts are kept at this time to allow flexibility in choosing
either tool. The geiger script is in `analyze_unsafe_code_geiger.py`.

`CalculateUnsafeCode.yml ` is updated to use the geiger script for
now.

Notes about the geiger script:

- The script runs cargo-geiger from each package directory to work
  around the repo being a virtual workspace.
- The shields.io badge JSON files created are organized by target
  directory.
- Markdown summaries are also created to show in the GitHub workflow
  UI.
- UEFI target compilcation is temporarily removed due to no_std
  compilation issues, so uses native target analysis is used instead.
  - This should be resolved in a future change.
- While unsafe code in dependencies is useful on its own, that data
  is excluded from these results which focuses on the unsafe code
  in packages in the repo.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>