Add Release Workflows (#10)
These two workflows are the building blocks for the updated way to
performing releases as specified in
https://github.com/OpenDevicePartnership/patina/issues/450.

The `CrateVersionUpdater.yml` is a reusable workflow expected to be
attached to a workflow that is manually triggered by a maintainer. This
workflow creates or updates a pull request to update the crate version
numbers of local crates to match the version specified in the current
draft release tag.

The updates to the `ReleaseWorkflow.yml` reusable workflow that
publishes a release to github and crates.io when a commit is pushed to
main that updates the crate version numbers to match the version
specified in the current draft release tag. It is expected that the
worfklow is executed on a push to main when one of the files updated is
"Cargo.toml".

This was tested via the following ways:

## CrateVersionUpdater

* verified to create a pull request when executed
* verified to update a pull request if one already exists
* verified to bail early if not executed on the main branch

## ReleaseWorkflow

* verified to bail early, but not fail the workflow if we are not on
main
* verified to bail early if the cargo.toml file update does not include
a version update
* verified to publish the correct release draft
* **NOT VERIFIED** to properly publish to crates.io