release worfklow: Bugfixes
This commit fixes two bugs in the release process. The first, associated
with `CrateVersionUpdater.yml` fixes an issue that causes the workflow
to fail in error if a release draft is not present. This choice was
purposeful as running the workflow when a release draft did not exist
seemed like an error. However this is a common occurance immediately
after a release has been performed. A new draft will not be created
until another PR merged.

The second, associated with `ReleaseWorkflow.yml` fixes a bug in which
I forgot to add `--no-confirm`, which is necessary for the publishing to
occur without human interference.