CiWorkflow.yml: Remove dry-release step (#23)
The dry-release step has caused more issues then it has solved and it is
not even able to test the main scenario that needs to be tested - when a
new crate is added. Due to this, I am removing the dry-release CI step.
If something does come up, then we can just create a PR to fix it and
perform another release.

Additionally - pre-set patina-devops version to v0.3.1