.sync/config.toml: Customize link section inclusion
Add a template parameter called `include_uefi_target_rules` to
control whether the linker arguments for UEFI targets are defined
for `rustflags`.

This is currently only needed for building UEFI binaries in the
patina-dxe-core-qemu and patina-readiness-tool repos.

PDB names can be customized for the `PDBALTPATH` parameter
independently for each target.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>