.sync/Makefile.toml: Clean Coverage Artifacts Before Running Coverage
Currently, this file is only synced to the patina repository. This
change brings in the changes from the following commit in the patina
repository to this copy of the file.

https://github.com/OpenDevicePartnership/patina/commit/faae91a29e096f7ecfffeaf414a3f823b1ea09f9

---

cargo-llvm-cov automatically cleans coverage artifacts before
running in order to have accurate coverage results. However, if
--no-report is passed, it does not automatically clean the coverage
artifacts.

Commit da004c7 added --no-report
to cargo-llvm-cov when refactoring the report output, but did not
add a call to cargo llvm-cov clean first. This is fine for CI builds,
but for local runs, it will produce inaccurate coverage if there are
existing coverage artifacts.

This fixes that by ensuring cargo llvm-cov clean is called first.

---

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>