Add all Patina repo Makefile.toml files for sync
Consolidate all Makefile.toml files as they exist in each repo to
be synced from this repo.

This makes it easy to audit files against each, leverage potential
shared implementation or templating in the future, and reduce
manual churn to update each makefile manually across repos when a
sweeping change is needed.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>